+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Git master
 - Use a password file and remove most ``no_log`` statements to aid in
   debugging.
 
+- Escape Groovy special characters in credentials (`#23_`).
+
 --------------------------
 Version 1.3.0 (2017-04-20)
 --------------------------
@@ -85,4 +87,5 @@ Version 1.0.0 (2016-06-29)
 
 .. _#14: https://github.com/solita/ansible-role-solita.jenkins/issues/14
 .. _#17: https://github.com/solita/ansible-role-solita.jenkins/issues/17
+.. _#23: https://github.com/solita/ansible-role-solita.jenkins/issues/23
 .. _#24: https://github.com/solita/ansible-role-solita.jenkins/pull/24